# REF/PERF: MultiIndex.get_locs to use boolean arrays internally